Add content IDs to unpublishing.
Switches the Unpublishing content_id to a fresh
one rather than that of the edition, and move the
edition content ID to the links has.

This allows the unpublishing to be a separate item
but still able to be replaced by the original
edition without having to delete the unpublishing
first.