Tag offsite consultations with topical events
Permits offsite consultations to be tagged with topical events. Prior to
this change, when checking the 'This consultation is held on another
website' box, the topical events field was hidden with a JS toggle. It's
now displayed whether or not the consultation is held offsite.

There is no validation logic around the value of the 'offsite' flag and
the presence of topical event tags, so this just works (tm). Persisting
consultations is well covered elsewhere.