Fix bug where unwithdrawing wouldn't update Panopticon
When invoked via Whitehall.edition_services, the change is broadcast to
listeners setup in config/initializers/edition_services.rb. This is wired up in
EditionServiceCoordinator with the 'notifier'. This code wasn't passing that
notifier in, so the listeners weren't being told about the change.

We think that this bug also means that unwithdrawn content wasn't being added to
Rummager, so this change fixes that too.