Don't index empty groups on collections pages
Document collections are pages on GOV.UK like this:

https://www.gov.uk/government/collections/future-of-cities

These collections have multiple "groups" of documents. When these
contain no documents they are hidden on the site.

However, the empty groups are still being sent to the search index.
This has caused some content to be unexpectedly indexed, like the above
mentioned page, which contains multiple empty groups with Lorem ipsum
placeholder text.

This commit changes the data sent to search, so that only content
visible on the page is indexed.

Thanks @tommorris for reporting a search result for "lorem ipsum".