Fix dependent deletion of Corporate Information Pages
When an object with Corporate Information Pages is deleted, we first
attempt to delete the CIPs. This was broken as it deleted each document
once for each edition of that document. The second iteration of this
loop would call `destroy` on `nil` as the document no longer exists.

Uniquifying the list of documents first before destroying them fixes
this issue.