Fix publishing_api integration
base_path is a required field with v2 of the publishing-api, but is
optional in the v1 endpoints.

govuk-content-schemas currently don't require base_path in the publisher
schemas, but we hope to address that by making v1 and v2 variants of the
schemas soon. This would enable us to make it required for v2 but still
optional for v1.

I found these test failures by making base_path required in my local copy of
govuk-content-schemas and then running the tests against them.