Unique path for worldwide org about pages.
The "about us" content for WorldwideOrganisations is displayed on the same page
as the organisation itself, which means we have two items with the same path.
This causes conflicts when the items are exported to publishing_api. For now,
fix this by giving the about us page its own slug, and always redirecting back
to the organisation home page. When these pages are properly migrated to
publishing-api, we will address the underlying issue.