Improve contextual nav for older browsers
We’re using a media min-width hack to apply some styles to browsers
that support media queries. These style use pseudo elements to create a
list item with a dash.

In older browsers, IE8 and below, users were seeing a numbered list,
which would render poorly when numbers got into double digits.

* Use a `disc` style rather than numbers to match intention of modern
browser styles
* Make the list position outside (the default), move the margin clause
outside of the media query as it now applies to both, and simplify the
list-style reset to only affect the type not the position.