Delete take part page controller
While the controller isn’t needed as Whitehall frontend is no longer
serving these routes, the `route` needs to be kept so that rails path
and url helpers are available, these are used in the Get Involved view,
and to generate the `search_link`.

Story to address this here:
https://trello.com/c/wvdvAKek/

Also removes assertions that the Take Part pages render correctly.