Remove unmaintained publishing-api populator code
This was once used to push all the content items from Whitehall to publishing-api.
This probably made a lot of sense in the early stages of
publishing-api/content-store, but nowadays its database is stable and supported.

The list of things is also incomplete. We use
lib/data_hygiene/publishing_api_republisher.rb for this purpose now