`Publication` publishing redirects `StatisticsAnnouncement`
This was previously done in the controller but has now been implemented
by simply updating (with touch on the association) the
`StatisticsAnnoucement` from the `Publication`.

This has involved making `PublishingApiPresenters::ItemRedirect` less
general (and consequently renamed to `StatisticsAnnouncementRedirect`
and the addition of some logic to
`PublishingApiPresenters.presenter_for` which doesn't really belong
there. There is an increasing need for a rethink/refactor around this
whole presenter workflow which was deemed beyond the scope of the
current task.