Correctly republish editions
Until now, it was easy to accidentally republish Editions in a way that could
publish the wrong content or lose the draft version from publishing-api.

This is because the implementation was originally written when the draft and
live content items were separate resource. That meant that you could
independently push live and draft items without them affecting each other.

Now, however, content is published by promoting it from draft to live, via the
publishing-api's publish action.

This means that the order of requests, when republishing the draft and live
editions of the same document, matters, because otherwise you could publish the
draft edition or overwrite the draft edition with the live edition.

I note that I have broken lots of calls to DataHygiene::PublishingApiRepublisher
that exist under db/data_migrations. There were already some examples
referencing PublishingApiEditionWorker, which hasn't existed for some time.