Bump gds-api-adapters to get better errors back
Since 27.1.0, gds-api-adapters includes the request body in the error message,
allowing you to see much more detail by default. This is particularly helpful
when trying to debug Whitehall pushing to the publishing-api.

I've bumped to the latest gds-api-adapters because there's no reason not to,
and it includes some useful changes.