Serve Govspeak images from CDN
Uses the `GOVUK_ASSET_HOST` environment variable instead of
`GOVUK_ASSET_ROOT` for the image host. This is needed because
`whitehall-backend` boxes are in charge of rendering migrated content
before sending to the publishing api, but
are configured to serve assets from `whitehall-admin.${app_domain}`.
That configuration was only intended for assets like CSS and JS that
are used in the Whitehall admin interface.