Add publishing-api sync check test
When provided with a format and a set of expectations, this
compares the data in the content-store with the persisted model in
Whitehall.

This is needed as when migrating formats, we have to republish all instances of the content
and we want to be sure they have been republished successfully.