Link to public classification page
When clicking “View on website” for classifications it shows the
whitehall admin rendered version of it.

Viewing a topical event page from `whitehall-admin. …` means the
relative link to its about page also points to `whitehall-admin`,
however that is now served by government front-end and this route
errors.

Classification changes go live immediately. Instead send users to the
public host with a cache bust query parameter.

Create a separate `public_classification_url` helper so as not to
interfere with:
* atom feed generation
* symmetry between existing helpers and topic_url/topical_event_url