Update attachment accessibility toggler markup
We want to migrate formats with attachments to Content Store, to be
rendered by government-frontend, where the govspeak is rendered by
the `govspeak` component from static.

To do this we need to port the styling and behaviour of attachments
to the `govspeak` component. This is easily done for SCSS, but the
JS that powers the "toggle request alternative format" is quite
WH specific, and it doens't make sense to port it to static,
especially as the current implementation has some accessibility
issues itself.

Instead, we're updating the attachment markup in the WH template to
use a new Module based toggler, that already exists on static and
is more accessible.

As WH formats being published to
the content store will push the Govspeak body, including the full
attachment markup rendered by `app/views/documents/_attachment.html.erb`
and making changes to that template will require republishing the
documents to content store, so we're making the change to the markup
in advance of publishing to content store.

The new markup/JS will continue to work within WH, as we migrate
formats to content-store / government-frontend.