Simplify inline attachment styling
Rather than sharing, and then having to reset, styling with the full
size attachment embed, give the inline version a different class,
which actually requires no styling, as it's just a link + text.

The ID on the span, previously provided by `content_tag_for`, is
depended on by the tests - and seemingly _only_ the tests.

I tried refactoring the tests, to remove the ID in the template,
but it wasn't straightforward - in particular switching to a URL
based selector assert, which works, but it's not possible to stub
the `url` field for `:file_attachment` to set distinct values when
dealing with multuple inline attachments.