Fix rendering of attachments with thumbs for pub-api
Without this change, when publishing Working Groups to the publishing-api, it
would throw an error when trying to process the Govspeak for attachments which
have a thumbnail URL of the form:

/government/uploads/system/uploads/attachment_data/file/12/

The error was:

ActionView::Template::Error: undefined method `user_signed_in?' for nil:NilClass
	from /usr/lib/rbenv/versions/2.2.3/lib/ruby/gems/2.2.0/gems/actionpack-4.2.5.2/lib/abstract_controller/helpers.rb:67:in `user_signed_in?'
	from /var/govuk/whitehall/app/helpers/application_helper.rb:289:in `path_to_image'
	from /usr/lib/rbenv/versions/2.2.3/lib/ruby/gems/2.2.0/gems/actionview-4.2.5.2/lib/action_view/helpers/asset_tag_helper.rb:211:in `image_tag'
	from /var/govuk/whitehall/app/helpers/document_helper.rb:136:in `attachment_thumbnail'
	from /var/govuk/whitehall/app/views/documents/_attachment.html.erb:20:in `block in _app_views_documents__attachment_html_erb___4417023685693832057_70194409418220'

This is because the `user_signed_in?` method is defined on the `controller`
which is nil when rendering views outside of a controller.

Digging into when this code was added, we couldn't see a reason to retain it
over using Rails' built-in `path_to_image`. This commit is effectively a revert
of 70798d63119a9ebd046fbaaf6289f4f195fb4336. All images are accessible via
either whitehall-admin.publishing.service.gov.uk or via the asset host.