Check draft content store sync with publishing api
With this change, for each base path and locale the publishing api sync check
makes one request to the content store, and another one to the draft content-store.

The check will generate a failure if the content item is missing from either content stores,
and the failure description will specify which content store(s) it is missing from.