Rescue from Addressable's exceptions
Since the upgrade to Ruby 2.2, the broken link checker has been
failing. This was due to a malformed link causing Addressable’s parsing
to fail. In this instance, we want to carry on with the rest of the
report but report this link as broken.

We can’t stub this link in Webmock as it uses Addressable under the
hood, but it’s enough to make sure that it wouldn’t be parsed and
ensure it gets returned as a broken link.