Remove published_at from consultation outcome attachments
The date an attachment is added to an outcome isn’t accurate - it gets
updated on every change to attachments. Similarly, using the published
date of the outcome itself is misleading for documents that are
published on different days. We should remove the date so as not to be
misleading to the public and policy teams.

Trello:
https://trello.com/c/gzLsPhhS/323-metadata-for-consultation-outcomes-doc
uments-displaying-incorrectly-small