Fix broken scheduling test
The `old_lock_version` in this test was actually referring to the
Edition ID. Most of the time this would be > 1, but now we’ve begun
using truncation in tests this  will always be 1, making the test fail.
This now uses the lock_version of the created edition as it should have
done. This should always be 0.