Track admin notices as Google Analytics events
After this change, any notice in Whitehall admin (eg "This document
has been withdrawn") will be sent to Google Analytics as an event.

This will allow us to count:

- how many times specific publishing transitions occur
- the frequency of errors encountered by publishers

The most immediate motivation for this work is from wanting to track how many times
unwithdrawing is done within GA (see #2497 for details).