Soft delete of attachments
Publishers can currently hard-delete HTML attachments, even those that were published with previous editions. When the edition with the deleted attachment is published, fetching by the attachment's slug starts returning 404, instead of redirecting or returning 410 gone. This results in broken journeys for users who follow those outdated attachment URLs.

The first step to rectifying this issue is to stop hard-deleting attachments; when HTML attachments are migrated to the single publishing stack, deletion of previously published attachments can result in a redirect item.