Remove composite unique index on `attachments`
There was a composite unique constraint on attachments composed of
`attachable_type`, `attachable_id` and `order` that had been added,
seemingly, to prevent attachments with duplicate `order` being added to
an `Attachable`.

Since implementing soft delete functionality in
https://github.com/alphagov/whitehall/pull/2518, violations of this
constraint have begun to occur causing exceptions.

This commit removes the constraint which appears to be redundant in any
case as the ordering is done via ajax drag and drop (so there is no
explicit user input to guard against) and then takes place within a
transaction [here](https://github.com/alphagov/whitehall/blob/master/app/models/attachable.rb#L117-L139)
so should be safe. The consequences of a duplicate order field are
minor in any case and the necessity of the soft delete functionality for
`HtmlAttachment` would outweigh this.