Only send links for published editions
Most items that contain links are descendents of Edition.

This makes the task behave the same regardless of whether you specify Edition
or one of its subclasses. We don't want to send links for drafts because if there
is already a live edition, with the current workflow publishers do not expect
changes to links to take effect immediately. This will change when we split the UI
for tagging.

There are some things in whitehall that are not editions, which may still need
republishing, but we'll do the editions first.