Change organisation presenter, update crests.
Organisations that have the `no-identity` or `custom`
OrganisationLogoTypes are currently publishing those class_names as
crests, which doesn't make sense for our frontends.

After discussing with @dsingleton and @fofr, we've decided to publish
`null` instead. It's preferable to work around it in whitehall than in our
frontends.

This will fix an issue in `government-frontend`, which is displaying the
"Environment Agency" crest incorrectly.

Relevant Trello ticket:
https://trello.com/c/cJGrtnc1/286-5-html-publications-migration-front-end-work-large