Use title terms index to find Editions
Replace the slow `LIKE` query with a query over the
`edition_title_terms` table to find just the IDs to filter the eventual
query by.

This splits the requested keywords up using the same method used to
populate the terms table, which is just a slug split by hyphen. It
first finds all editions with any of the terms in the title. Then it
asks MySQL to join the sorted terms found with commas, and filters only
by those records having the exact same terms as those requested.

Arel needs to be used as ActiveRecord’s abstracted query interface
doesn’t support the complex query needed. The table also doesn’t have
an associated model.

The original slug query has also been switched out for something that
fetches just the ID for that slug, which can then be passed to an
`IN()` filter for the actual  editions, along with the IDs from the
title query.

This brings a title search down from around 30 seconds (and a likely
timeout) to under a second.