Publish intents for StatisticsAnnouncement.
This adds additional callbacks when creating, updating, or deleting
a `StatisticsAnnouncement` to send `publish_intents` to the publishing
API.

This is necessary to set the cache expiration correctly for the
publication of the statistics themselves.

This functionality is only exposed on the `v1` `publishing_api`.
Ultimately, this will happen internally, within the `publishing_api`, so
using the `v1` endpoint seems like a reasonable temporary solution.

Relevant Trello ticket:
https://trello.com/c/zMdu3dHe/343-use-publish-intents-on-statistics-announcements-medium