Publish placeholders for static pages
Currently the static pages are not present in the new publishing platform and
do not have content ids. This ensures they are present with placeholders.

We will migrate the links once, so that any tags from content tagger do not get
overriden in the future when we deploy.

The rake task that publishes to rummager and publishing api is currently manually
run - we'll rerun it for this deploy.