Make GovspeakRenderer threadsafe
The reference to `ApplicationController.helpers` was a singleton
instance of `ActionView::Base`. This class is used to contain the
context for rendering a single view at a time.

Instead, we need a new instance of `ActionView::Base` for each instance
of the `GovspeakRenderer`. We also need to extend this instance with
all the helper methods that would normally be part of the view context.

This fixes an issue with multiple govspeak bodies being rendered by
different threads at the same time having corrupted output.

Additionally, the `helpers` private method is now called `view_context`
in order to more accurately signify it’s use.