Update links presentation in publishing API presenters
Add a '#links' method to each presenter, replacing the '#filter_links'
convention. This:

- allows for flexibility in determining how links are presented,
  decoupling this behaviour from the implementation in
  PublishingApiPresenters::LinksPresenter.
- makes it a little clearer and more obvious how links are set in a
  given presenter.

Taking this approach allows the LinksPresenter to be used as a sensible
default but permits, for example - setting multiple links in the links
hash based on several content IDs retrieved from the publishing API in a
single network call. This specific use case will be required in a
followup commit that changes how Edition links are presented.