Explain current biography logic to users
The current process for creating or editing people is confusing. We
have some logic that shows a truncated biography for any person without
a current role. If you’re creating a new person, they have no role –
when you enter your full biography and check it on the live site, it’s
not clear why the whole thing isn’t showing.

This is a simple change to try and explain that. The decision to do the
truncation should be re-assessed.

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/1298476
Original PR to add truncation:
https://github.com/alphagov/whitehall/pull/2058