Refactor `GovspeakContent` and `GovspeakContentWorker`
This commit moves all of the govspeak -> html rendering logic into the
`GovspeakContent` class. This will allow it to be called elsewhere e.g. when
the `HtmlAttachment` is being presented for the Publishing API. When
that is done we may be able to get rid of the `GovspeakContentWorker`
altogether which I'll do in a further PR if possible.