Exclude `cancelled?` and `unpublished?`
Running this task previously was causing a lot of `500` responses from
the publishing stack due to the way `404` responses returned by the
`publish_intents` endpoint on Content Store are handled on their way back
through the stack. This change stops `unpublished?` and `cancelled?`
`StatisticsAnnouncement` being sent. These states cause a delete publish
intent call to be made resulting in the error when it isn't found on the
content store.