Call `GovspeakContent#render_govspeak!` in the presenter
The conversion from Govspeak to HTML is currently done asynchronously in
`GovspeakContentWorker` which means that the HTML may not be present
when the `HtmlAttachment` is presented. Calling it here ensures that it
is and as this presenter will be used within one of the
`PublishingApiWorker` processes any performance hit won't block.