Allow for missing primary specialist sectors
We've run into cases where the whitehall database refers to
non-existing primary specialist sectors. This causes our PATCH links
call to publishing-api to fail.

We've considered fixing this data in the database, but decided against
it because the data in the database will be removed soon. Updating the
records in the database now would require manual intervention and could
have unforeseen side-effects.