Add related_mainstream links and details to DetailedGuide presenter.
DetailedGuides have `related_mainstream_content_url` and
`additional_related_mainstream_content_url`. This commit updates the
presenter to fetch the content items from `PublishingApiV2` for these
items, and serialize their IDs as links for the `ContentApi` to
substitute for the real items later.

Additionally, the order is preserved using
`details.related_mainstream_content`, a sorted array of the
aforementioned IDs.