Add a data migration to republish all DetailedGuides
Whitehall currently has ~1700 published detailed guides that do not have
a corresponding content item in the content store.

We recently changed frontend behaviour for documents rendered by
Whitehall - specifically, parts of the header now rely on the content
item to pull in correct tagging information.

This data migration uses the existing #republish_document_async method
to send both the published edition and the pre-publication edition (if
present) to the V2 publishing API.