Don’t send request_id if the job pre-dates request tracing…
For jobs that were enqueued before request tracing
was added, we need to make sure we clear the
govuk_request_id header. This is because
GdsApi::GovukHeaders contains global state that
will hang around in the Sidekiq worker after the
job has run.

See the discussion on the pull request for more
context: https://github.com/alphagov/whitehall/pull/2567