Split format into schema_name and document_type
We now have split schema_name and document_type fields in
publishing-api to replace the single format field. This is used to
model documents that have a shared schema but are different types.

For the most part, both fields will be the same, which is how
publishing-api split the format by default. But for placeholders, stats
announcements and unpublishings we can now model the document with more
granularity.

This is needed for documents linked from HTML Publications, as they
need to know the type of their parent document. This is currently just
‘placeholder’ for the most part, so we need to have an accurate
document_type set.