Respect soft delete for attachment visibility
When checking whether an attachment should be visible for `Edition`
based content we need to check that it has not been soft-deleted ie has
`deleted? == true`.

For `Editions` the associated `AttachmentData` object may not get
deleted when the `FileAttachment` itself is deleted if it is also
associated with a previous `Edition`s `FileAttachment`.

For non-`Edition` content no check is required as the `AttachmentData`
object always gets deleted as there are no prior versions of the
content for it to be related to.