Present national_applicability for detailed guides
Some detailed guides have `.nation_inapplicabilities`, which is a
collection of `Nation`s that this guide does not apply to.

This will present a `national_applicability` object which specifies
which countries are not applicable, and if available, an alternative url
to find applicable content.

The relevant schema for this is in
https://github.com/alphagov/govuk-content-schemas/pull/300.