Use locale to build base_path in presenters
Whitehall’s publishing api workers use the I18n with_locale block to
tell presenters which locale to present for. This needs to be passed
through to the url helpers in order to append the locale on the
base_path.

This uncovered a bug in how we call url helpers which was ignoring the
requested locale if `routing_type: :path` was set. The correct
invocation is `only_path: true`.