Remove detailed guides API.
Nobody uses it but some bots and the odd lonely
soul in search of guidance in JSON form.

This will be replaced by the content store once
detailed guides have been migrated, though if we
want to have the rel=alternate URLs in the page
we should think about putting that in across the
board.