Add some extra documentation
While debugging an issue with statistics announcements yesterday we
learned a bunch of things about how Whitehall works. This commit is an
attempt to put some of that back into the code as comments. The idea is
that these comments would’ve helped us yesterday, so it might help
someone the next time. It’s by no means complete but hopefully it’s a
small improvement.