More sidekiq workers in production
We're now doing bulk republishing of documents from Whitehall, some of which take some time to process. The jobs are a mix of CPU and publishing-api bound. Adding more workers will improve throughput. There is enough space in CPU and RAM on the Whitehall backend boxes. This has been verified in integration.