Use the bulk_republishing queue in PubApiDocRepublisher
This is a special republisher that is used for Editions, but it seems
to be currently using the `default` queue. This means that when it's
used it will flood the `default` queue and block all other
republishing until it's done.

This commit switches it to use the `bulk_republishing` queue.