Send contacts to the publishing api...
Present contacts in the correct format and send to the publishing api.
Contacts do not have a base path or routes so their presentation doesn't
follow the standard ItemPresenter inheritance pattern.
The contact links should contain the country and organisation of the contact.
The details are comprised of email addresses, phone numbers and postal addresses.