Add explicit publishing API presenters for 4 document types
This commit adds presenters for the Person, MinisterialRole,
WorldLocation, and WordlwideOrganisation items.

Why?

These were previously presented by the publishing API
PlaceholderPresenter. This presenter was attempting to extract topic and
organisation link types from the item, regardless of whether the item
could be tagged to topics and organisations. By providing specific
presenters for each item, we gain the following:

- We can specify the links payload explicitly rather than relying on the
  implementation in PlaceholderPresenter.
- It's much clearer how items of this type are presented to the
  publishing API.

How?

- Add new presenter classes for each of the 4 items.
- Move the pre-existing test cases from PlaceholderTest into individual
  test files.
- Update the PublishingApiPresenters case statement to select the
  appropriate presenter at runtime.