Remove PlaceholderPresenter
As part of an effort to simplify the structure of the publishing API
presenters, this commit removes the PlaceholderPresenter and updates its
dependents so that these can continue sending a correct payload to the
publishing API.

Why?

We're moving away from using inheritance as a mechanism for sharing
functionality among the publishing API presenters. The
PlaceholderPresenter was a parent class of the Organisation,
PolicyAreaPlaceholder, and TopicalEvent presenters. This relationship
created a tight coupling between those child presenters and the
Placeholder parent. Like other publishing API presenters, these are now
more responsible for determining their own payloads.

They still inherit from PublishingApiPresenters::Item, however this
relationship will also be removed in a subsequent commit.

How?:

- Remove PlaceholderPresenter from the PublishingApiPresenters resolver
  module.
- Raise an error in the resolver if no presenters match the input item.
- Move a locale-related test from PlaceholderTest to ItemTest.
- Delete the PlaceholderPresenter file and associated test.
- Add missing PolicyAreaPlaceholderTest.
- Add additional methods to presenter classes that were previously
  inheriting from PublishingApiPresenters::Placeholder.