Refactor a subset of Edition presenters
- Stop inheriting from EditionPresenter in CaseStudy, DetailedGuide,
  Publication, and DocumentCollectionPlaceholder presenters. These
  presenters were chosen because they are reasonably isolated from
  other presenters in this subsystem. The remaining presenters will be
  updated in a follow-up change.
- Extract common functionality into a set of PayloadBuilder objects that
  return specific parts of the publishing API payload for a given item.
  This allows each presenter to compose its complete payload using one or
  more of these objects, as well as attributes defined locally.
- Introduce a BaseItem presenter that provides attributes common across all
  documents sent to the publishing API.
- Extract default_update_type into a mixin.

Why?:

Inheritance isn't the right choice for sharing code amongst these
presenters. The children are not really proper subtypes of
EditionPresenter or ItemPresenter. This, coupled with the 3-level class
hierarchy made this part of the codebase hard to understand and
difficult to change. By adopting a composition-based design instead, the
contents of a given publishing API payload are easier to determine and
modify.