Fix module/class definitions in PublishingApiPresenters
These were being defined in the form `class ModuleName::ClassName`. This
is incorrect - it isn't syntactially the same as nesting the class under
the module, results in a different return value for Module.nesting in
the innermost class, and therefore leads to unexpected /
counter-intuitive behaviour when accessing constants.

This commit also removes the following:

- The explicit app/* entries in autoload_paths. We don't need these
  because Rails adds app/* to autoload_paths by default.
- Requiring of the PublishingApiPresenters top-level module. This
  shouldn't be necessary in any of these classes.