Only require active model serializers where needed
By allowing the app to require the gem in the Gemfile, we will see side effects
in the controllers as the middleware will try to find serializers to use
for the classes in every `render: json` call. This way, we only use active
model serializers in the case study presenter and in the tests.