Dynamic Prime Minister on "How government works"
This commit changes one of the more static page to load the current
Prime Minister from the database. This means that we don't have to
manually update it each time the Prime Minister changes.

Some unrelated tests used to use the 'How government works' page
(presumably because it didn't require anything to be in the database).
I've changed a couple of those references to more simple pages.