Order cabinet members by cabinet role seniority
Two cabinet ministers are appearing on the ministers page
(https://www.gov.uk/government/ministers) in the wrong order.

Cabinet ministers are appearing on the ministers page according to
their role seniority, regardless of whether the role is a cabinet role
or not.

This means that cabinet ministers that also have another role, might
not be presented in the frontend in the same order as the "cabinet
members ordering" list in Whitehall admin.

This ensures that only the seniority of cabinet roles are taken into
account.

Zendesk ticket: https://govuk.zendesk.com/agent/tickets/1378654