Update SpecialistTagFinder to use request path for content store call
SpecialistTagFinder needs to fetch the current document from the content
store in order to retrieve its links data. These links are used when
rendering the document header. Previously the base path of the current
document was derived via the PublishingApiPresenters. This has proven to
be fragile as the @document being rendered on the page may be an
instance of a given model, or an instance of the model's corresponding frontend
presenter.

Following a recent refactor, the PublishingApiPresenters
'.presenter_for' method no longer provides a convenient catchall for
anything that doesn't match one of the classes in its case statement.
This makes the model vs. publishing presenter relationship more
transparent, but has lead to errors when an instance of a frontend
presenter gets passed in.

By using the current request path instead, we decouple
SpecialistTagFinder from the PublishingApiPresenters and avoid this
issue.