Add Group#update_memberships_and_attributes method
We want to preserve the order of members on updating a group. This
method is an atomic action that destroys all memberships and then
rebuilds them before updating the model. This allows us to display a member
order on the front end relying on the join model id order to reflect the
order that the members were added on the admin form.