Add search parameter to display the finder as open
The finder on the search results page is normally shown closed, unless at least one filter has been selected, in which case it is shown open. This adds a search parameter that forces the finder to be shown open even if no filters are selected, for all searches originating from Whitehall.

The substantive functionality is implemented in Frontend (see https://github.com/alphagov/frontend/pull/984).