`PublishingApiGoneWorker` supports `alernative_url` and `explanation`
When a document is 'unpublished, published in error' the user can supply
optional alternative url and explanation values and choose whether the
document should redirect automatically or just display the alernative
url and explanation.

This commit adds support for that to the `PubishingApiGoneWorker`.
Publishing API alreadys sends these values in the `Gone` payload to
Content Store.