Pass `Document#id` instead of edition ids to republishing worker
The `PublishingApiDocumentRepublisher` currently accepts
`published_edition_id` and `pre_publication_edition_id` arguments. This
means the editions have to be retrieved from the DB prior to queueing
and then retrieved again within the worker when the job is run.

This commit updates the worker to accept the `Document#id` instead,
halving the number of DB calls required.