Fix “invalid specialist sector” error in Whitehall admin
Currently, when creating a new document in Whitehall admin, if any combination of primary and/or secondary specialist sector tags are specified then a “specialist sectors are invalid” error message is displayed. Updating a document to add such tags works as expected.

This is due to model validation to ensure any specific tag is not specified more than once for the same document. Due to existing tags in the `specialist_sectors` table that have `edition_id` set to `null`, and the fact that `edition_id` is `null` for a new document which has not yet been saved, most tags are seen as already existing for a new document, hence the validation fails.

This commit does the following:

* Deletes all rows in the `specialist_sectors` table that have `edition_id` set to `null`. These rows are not tagging any particular document and are therefore not required.

* Makes the `edition_id` and `tag` columns in the `specialist_sectors` table not-nullable to ensure such records cannot be re-inserted later.