Rename Document CSV header column
'Affected by history-mode' -> 'Can have history-mode'

This is a better name as the column data is dependant on whether or not
the document is "political" or not, and therefore *can* have history mode
applied. A later commit will add data on whether the history mode
is actually applied or not.