`PublishingApiGoneWorker` supports `allow_draft`
This makes it possible to pass the `allow_draft` argument through to the
`/unpublish` endpoint of Publishing API. We need to be able to do this
to allow unpublishing of things without publishing them first when
bulk republishing during migration. Don't worry... I wrote that last
sentence and I don't understand it either.