Add `LinksCheck`
This check accepts an array of `content_id` UUIDs that should be present
at a key in the links of the content store representation of the item.
It returns an error if there are missing content ids, unexpected content
ids or if the key is missing altogether from the links.