Add `TopicsCheck`
So... the `Topic` in question are actually `specialist_sector_tags` and
getting a `content_id` for these requires a call out to PublishingApi.
Implementing this check as a `LinksCheck` would have meant having all of
that and the logic that goes with it in the format specific checks so
I've implemented it as a separate type.

There is duplication with `LinksPresenter` but I think it is reasonable
as this is 'test' code and we have been trying to avoid calling the
presenters withint the test as they are part of the stack we are
testing.

This check also checks `parent` which is an array of one link and is
also actually a `specialist_sector_tag` in Whitehall. This avoids having
two checks making API calls. For to make it go a bit faster an' all
that.