Change all instances of unified_search to search
This commit changes all instances of unified_search to search when accessing rummager. This provides consistency between the internal and external APIs in terms of naming.

As a side effect of upgrading the version of gds-api-adapters, this commit also changes `ci_reporter` to `ci_reporter_minitest` since version 2 of this gem has been split for different test frameworks.